# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Manga {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImage?: string | null;
  type: string;
  status: string;
  genres?: string[];
  chapterCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface MangaInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  coverImage?: string;
  type: string;
  status: string;
  genres?: string[];
}

export interface MangaUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  coverImage?: string;
  type?: string;
  status?: string;
  genres?: string[];
}

export interface MangaListResponse {
  data: Manga[];
  total: number;
  page: number;
  limit: number;
}

export interface MangaWithLatestChapter {
  id: number;
  title: string;
  /** @nullable */
  coverImage?: string | null;
  type: string;
  status?: string;
  /** @nullable */
  latestChapterNumber?: number | null;
  /** @nullable */
  latestChapterTitle?: string | null;
  /** @nullable */
  latestChapterId?: number | null;
  updatedAt?: string;
}

export interface Chapter {
  id: number;
  mangaId: number;
  number: number;
  /** @nullable */
  title?: string | null;
  status: string;
  pageCount?: number;
  createdAt: string;
  /** @nullable */
  publishedAt?: string | null;
}

export interface ChapterInput {
  number: number;
  title?: string;
}

export interface Page {
  id: number;
  chapterId: number;
  pageNumber: number;
  imageUrl: string;
  createdAt?: string;
}

export interface ChapterWithPages {
  id: number;
  mangaId: number;
  number: number;
  /** @nullable */
  title?: string | null;
  status: string;
  createdAt?: string;
  /** @nullable */
  publishedAt?: string | null;
  pages: Page[];
  manga?: Manga;
}

export interface ChapterWithManga {
  id: number;
  mangaId: number;
  number: number;
  /** @nullable */
  title?: string | null;
  status: string;
  pageCount?: number;
  createdAt: string;
  manga?: Manga;
}

export interface PageReorderInput {
  pageIds: number[];
}

export interface PublisherAuthInput {
  code: string;
}

export interface PublisherAuthResult {
  success: boolean;
  token: string;
  /** @nullable */
  message?: string | null;
}

export interface SiteStats {
  totalManga: number;
  totalChapters: number;
  totalPages: number;
  publishedChapters?: number;
  pendingChapters?: number;
}

export interface RemoteImportInput {
  url: string;
  mangaId: number;
  chapterId: number;
}

export interface RemoteImportResult {
  success: boolean;
  importedPages: number;
  /** @nullable */
  message?: string | null;
  errors?: string[];
}

export interface RemotePreviewInput {
  url: string;
}

export interface RemotePreviewResult {
  title: string;
  pageCount: number;
  pageUrls: string[];
  /** @nullable */
  chapterNumber?: number | null;
  /** @nullable */
  mangaTitle?: string | null;
}

export interface UserRegisterInput {
  /** @minLength 3 */
  username: string;
  email?: string;
  /** @minLength 6 */
  password: string;
  avatar?: string;
}

export interface UserLoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResult {
  success: boolean;
  token: string;
  /** @nullable */
  message?: string | null;
  user: User;
}

export interface Comment {
  id: number;
  mangaId: number;
  userId: number;
  content: string;
  createdAt: string;
  user?: User;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export type ListMangaParams = {
search?: string;
type?: string;
status?: string;
genre?: string;
page?: number;
limit?: number;
};