# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RTN Manga API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manga/{mangaId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters for a manga
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create chapter (publisher only)
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /manga/{mangaId}/chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get chapter with pages
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Chapter with pages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterWithPages"
        "404":
          description: Not found
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter (publisher only)
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /manga/{mangaId}/chapters/{chapterId}/publish:
    post:
      operationId: publishChapter
      tags: [chapters]
      summary: Publish a pending chapter
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: Registered user with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error or username taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Login result with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /comments/manga/{mangaId}:
    get:
      operationId: getMangaComments
      tags: [comments]
      summary: Get comments for a manga
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addMangaComment
      tags: [comments]
      summary: Add a comment to a manga
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (author only)
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Manga:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        genres: { type: array, items: { type: string } }
        chapterCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Chapter:
      type: object
      required: [id, mangaId, number, status, createdAt]
      properties:
        id: { type: integer }
        mangaId: { type: integer }
        number: { type: number }
        title: { type: ["string", "null"] }
        status: { type: string }
        pageCount: { type: integer }
        createdAt: { type: string }
        publishedAt: { type: ["string", "null"] }
    ChapterInput:
      type: object
      required: [number]
      properties:
        number: { type: number }
        title: { type: string }
    ChapterWithPages:
      type: object
      required: [id, mangaId, number, status, pages]
      properties:
        id: { type: integer }
        mangaId: { type: integer }
        number: { type: number }
        title: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        publishedAt: { type: ["string", "null"] }
        pages:
          type: array
          items:
            $ref: "#/components/schemas/Page"
        manga:
          $ref: "#/components/schemas/Manga"
...
    Page:
      type: object
      required: [id, chapterId, pageNumber, imageUrl]
      properties:
        id: { type: integer }
        chapterId: { type: integer }
        pageNumber: { type: integer }
        imageUrl: { type: string }
        createdAt: { type: string }
...
    UserRegisterInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        avatar: { type: string }
    UserLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: ["string", "null"] }
        avatar: { type: ["string", "null"] }
        createdAt: { type: string }
    AuthResult:
      type: object
      required: [success, token, user]
      properties:
        success: { type: boolean }
        token: { type: string }
        message: { type: ["string", "null"] }
        user:
          $ref: "#/components/schemas/User"
    Comment:
      type: object
      required: [id, mangaId, userId, content, createdAt]
      properties:
        id: { type: integer }
        mangaId: { type: integer }
        userId: { type: integer }
        content: { type: string }
        createdAt: { type: string }
        user:
          $ref: "#/components/schemas/User"
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }