# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResult {
  success: boolean;
  token: string;
  /** @nullable */
  message?: string | null;
  user: User;
}