# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Chapter {
  id: number;
  mangaId: number;
  number: number;
  /** @nullable */
  title?: string | null;
  status: string;
  pageCount?: number;
  createdAt: string;
  /** @nullable */
  publishedAt?: string | null;
}