# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Comment {
  id: number;
  mangaId: number;
  userId: number;
  content: string;
  createdAt: string;
  user?: User;
}