# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegisterInput {
  /** @minLength 3 */
  username: string;
  email?: string;
  /** @minLength 6 */
  password: string;
  avatar?: string;
}